requirements.txt
torch>=1.9.0
torchvision
regex
ftfy
tqdm
huggingface_hub
datasets 

pillow
gdown
open-clip-torch